+++ src/services/breakpoint.test.ts
@@ -21,7 +21,16 @@ import {
   getBreakpoint,
   EuiBreakpoints,
   isWithinMaxBreakpoint,
+  isWithinBreakpoints,
 } from './breakpoint';
+
+const CUSTOM_BREAKPOINTS: EuiBreakpoints = {
+  xl: 1400,
+  l: 1078,
+  m: 812,
+  s: 667,
+  xs: 320,
+};
 
 describe('getBreakpoint', () => {
   describe('with default BREAKPOINTS', () => {
@@ -41,14 +50,6 @@ describe('getBreakpoint', () => {
       expect(getBreakpoint(1400)).toBe('xl');
     });
   });
-
-  const CUSTOM_BREAKPOINTS: EuiBreakpoints = {
-    xl: 1400,
-    l: 1078,
-    m: 812,
-    s: 667,
-    xs: 320,
-  };
 
   describe('with custom breakpoints', () => {
     it("should return 'undefined' for 240", () => {
@@ -91,14 +92,6 @@ describe('isWithinMaxBreakpoint', () => {
     });
   });
 
-  const CUSTOM_BREAKPOINTS: EuiBreakpoints = {
-    xl: 1400,
-    l: 1078,
-    m: 812,
-    s: 667,
-    xs: 320,
-  };
-
   describe('with custom breakpoints', () => {
     it("should return 'false' for 'xs' and 240", () => {
       expect(isWithinMaxBreakpoint(240, 'xs', CUSTOM_BREAKPOINTS)).toBe(false);
@@ -123,6 +116,57 @@ describe('isWithinMaxBreakpoint', () => {
   describe('with max as a number', () => {
     it("should return 'true' for a 320 width and a '480' max", () => {
       expect(isWithinMaxBreakpoint(320, 480)).toBe(true);
+    });
+  });
+
+  describe('isWithinBreakpoints', () => {
+    describe('with default BREAKPOINTS', () => {
+      it("should return 'true' for 'xs' and 320", () => {
+        expect(isWithinBreakpoints(320, ['xs'])).toBe(true);
+      });
+      it("should return 'true' for 's' and 667", () => {
+        expect(isWithinBreakpoints(667, ['s'])).toBe(true);
+      });
+      it("should return 'true' for 'm' and 812", () => {
+        expect(isWithinBreakpoints(812, ['m'])).toBe(true);
+      });
+      it("should return 'true' for 'l' and 1078", () => {
+        expect(isWithinBreakpoints(1078, ['l'])).toBe(true);
+      });
+      it("should return 'true' for 'xl' and 1400", () => {
+        expect(isWithinBreakpoints(1400, ['xl'])).toBe(true);
+      });
+    });
+
+    describe('with custom breakpoints', () => {
+      it("should return 'false' for 'xs' and 240", () => {
+        expect(isWithinBreakpoints(240, ['xs'], CUSTOM_BREAKPOINTS)).toBe(
+          false
+        );
+      });
+      it("should return 'true' for 'xs' and 575", () => {
+        expect(isWithinBreakpoints(575, ['xs'], CUSTOM_BREAKPOINTS)).toBe(true);
+      });
+      it("should return 'true' for 's' and 768", () => {
+        expect(isWithinBreakpoints(768, ['s'], CUSTOM_BREAKPOINTS)).toBe(true);
+      });
+      it("should return 'true' for 'm' and 992", () => {
+        expect(isWithinBreakpoints(992, ['m'], CUSTOM_BREAKPOINTS)).toBe(true);
+      });
+      it("should return 'true' for 'l' and 1200", () => {
+        expect(isWithinBreakpoints(1200, ['l'], CUSTOM_BREAKPOINTS)).toBe(true);
+      });
+      it("should return 'true' for 'xl' and 1400", () => {
+        expect(isWithinBreakpoints(1400, ['xl'], CUSTOM_BREAKPOINTS)).toBe(
+          true
+        );
+      });
+    });
+
+    describe('with multiple sizes', () => {
+      it("should return 'true' for a 667 width and ['xs', 's']", () => {
+        expect(isWithinBreakpoints(667, ['xs', 's'])).toBe(true);
+      });
     });
   });
 });