+++ src/services/breakpoint.ts
@@ -79,3 +79,22 @@ export function isWithinMaxBreakpoint(
       : false;
   }
 }
+
+/**
+ * Given the current `width` and an array of breakpoint keys,
+ * this function returns true or false if the `width` falls within
+ * any of the named breakpoints
+ *
+ * @param {number} width Can either be the full window width or any width
+ * @param {EuiBreakpointSize[]} sizes An array of named breakpoints
+ * @param {EuiBreakpoints} breakpoints An object with keys for sizing and values for minimum width
+ * @returns {boolean} Returns `true` if current breakpoint name is included in `sizes`
+ */
+export function isWithinBreakpoints(
+  width: number,
+  sizes: EuiBreakpointSize[],
+  breakpoints: EuiBreakpoints = BREAKPOINTS
+): boolean {
+  const currentBreakpoint = getBreakpoint(width, breakpoints);
+  return currentBreakpoint ? sizes.includes(currentBreakpoint) : false;
+}